Extract `FileExists` tests into separate test case
